Fix the issue "Please remove functionality with default link configurations"
Per default the system loads a file named "default". Problem is, if the modelUrl
in this file differs from current selected model, the system crashes. Because of
that it was not possible to use CubeViz with other knowledge bases as scoreboard
Now, there is no default file anymore so it should be possible to use other
knowledge bases.

Temporary hack: If dsd, dataset or dimension or measure label contains
whitespaces, than they will replaced by an _