Change the key-naming for components (dim, meas): from label to hashedUrl
It was necessary because if you use white spaces in your label
it can kill your object-access, because object key's ought'nt
contain white spaces.

Another big change is the erease of the property_ and dimension_ stuff.
It was inconsistent to the rest of the system. Now its label for the
label of a certain dimension, typeUrl for type etc. pp.