Bunch of massive bugfixes: JSON transfer, changes for new data, ...
# Incorrect JSON transfer #
For some reason jQueries JSON encoding
did not work properly to transfer the massive objects containing
all the information about the dimensions and their elements.
There was always a gap in the data. To fix this, we stringify
the whole content (its a big object) with JSON.stringify. After
that everything works fine.

# Additional changes for new data #
Because we use all the information what the backend provides
directly (predicate as key, object as value) it is neccessary
to adapt additional code. There is no property, propertyLabel
or hashedProperty anymore! From now on there is something like
__cv_uri or the predicate/object pairs of the backend.
For instance: attach element direct to each checkbox in component
dimension configuration dialog for later usage. Or fixing sparql
in Query.php to get observations.

# Fix link generation in ViewHelper #