Fix problem of wrong deactivated select buttons and increase attr handling
Sometimes it happens that the select-all/deselect-all buttons
were deactivated. This was because of the missing update of the
new number of selected elements.

Increase attribute uri handling. It is now possible that the
attribute uri is not the predicate associated with an observation
but could also but the object with predicate "cube:attribute".