Fix massive bug: fix placement of values in series array
Problem was, up to this point CubeViz simply add new items
to the series array. But if there are elements with no
associated observation value, the number of the series
object data arrays differs which leads to a corrupt
chart.

To avoid that, each new item for object.data in series
gets initialized with as many null-value entries as
categories are in use. Additionally each new value for
series.obj.data gets placed based on the associated element
in categories list.