Further improvements for #173 (Using Object- and QueryCache)
Now CubeViz uses a more aggressive way to store
results in cache which means, that not only sparql
results but as well computations will also be stored.

Furthermore, retrievedObservations field was moved
backend back to data and is now stored together with
all the configuration which leads to a certain
list of retrieved observations.
This change will increase the performance, because there
is no getObservations-call to the data store everytime
the chart configuration gets updated.