Switch to use only local data (only dimensions yet) if using a merged dataset
Main change was to adapt View/DataselectionModule/Component.ts
that if you use a merged dataset, CubeViz will NOT pull further
information from the store, like new observations.

The reason is, that the merged dataset is only located in the
according dataHash-file (in ObjectCache) and any queries to store
will result in non-sense.

But CubeViz will recognize changes in Dimension/Dimension-
Element-Selection and marks all observations as active whose
fit with selected Dimension Elements.