wrong Query limit results in incomplete observations
setLimit on Query object limits the number of triples,
but not the limit of the elements itself!

so removing it and add a hard coded limit in the result
handling shall fix the problem. but from now on, we have
to load ALL according observations, which may result, in
some cases, into a higher load, but in the end all retrieved
elements are complete.